refactor: components 폴더 정리 및 필요없는 컴포넌트 제거
