+++ src/components/landing/GuideSection/GuideSection.tsx
