+++ src/components/landing/LandingMainSection/landingMainSection.css.ts
