+++ src/components/landing/StartSection/StartSection.tsx
