+++ src/components/landing/StudyEffectGuide/StudyEffectGuide.tsx
