+++ src/components/landing/StudyEffectGuide/studyEffectGuide.css.ts
