+++ src/components/landing/StudyStepGuide/StudyStepGuide.tsx
